flask_templates_databaseforuser/requirements.txt
webbrowser
jsonify
flask_cors
flask_wtf
wtforms
python-dotenv
flask
sqlalchemy
waiting